% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{coef.glmMixture}
\alias{coef.glmMixture}
\title{Extract Coefficients for glmMixture}
\usage{
\method{coef}{glmMixture}(object, ...)
}
\arguments{
\item{object}{A fitted glmMixture object.}

\item{...}{Additional arguments.}
}
\description{
Extract Coefficients for glmMixture
}
