% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.coxph_mixture.R
\name{confint.coxph_mixture}
\alias{confint.coxph_mixture}
\title{Confidence intervals from a `coxph_mixture` Object.}
\usage{
\method{confint}{coxph_mixture}(object, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class "coxph_mixture".}

\item{level}{the required confidence level.}

\item{...}{for additional `confint` arguments.}
}
\value{
a data frame with the upper and lower confidence limits for each
outcome model coefficient. These limits are labelled as (1-level)/2 and
1 - (1-level)/2 (e.g., 2.5\% and 97.5\% for level 0.95).
}
\description{
Compute confidence intervals for the outcome model coefficients.
}
