% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_glm_methods.R
\name{confint.glmELE}
\alias{confint.glmELE}
\title{Confidence Intervals for \code{glmELE} Objects}
\usage{
\method{confint}{glmELE}(object, parm, level = 0.95, weight.matrix = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"glmELE"}.}

\item{parm}{A specification of which parameters are to be given confidence intervals,
either a vector of numbers or names. If missing, all parameters are considered.}

\item{level}{The confidence level required.}

\item{weight.matrix}{Character string specifying the weighting method to use.
Defaults to the first method found.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each parameter.
}
\description{
Computes Wald confidence intervals for one or more parameters in a \code{glmELE} object.
}
\examples{
data(brfss, package = "postlink")

adj_object <- adjELE(linked.data = brfss,
                     m.rate = unique(brfss$m.rate),
                     blocks = imonth,
                     weight.matrix = "BLUE")

fit <- plglm(Weight ~ Height + Physhlth + Menthlth + Exerany,
             family = "gaussian", adjustment = adj_object)

confint(fit)

}
