% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.spline_mixture.R
\name{confint.spline_mixture}
\alias{confint.spline_mixture}
\title{Confidence Intervals from a `spline_mixture` Object}
\usage{
\method{confint}{spline_mixture}(x, level = 0.95)
}
\arguments{
\item{x}{the result of a call to `spline_mixture()`}

\item{level}{the desired confidence level (default is 0.95)}

\item{...}{for additional confint arguments}
}
\value{
a matrix with the confidence intervals
}
\description{
Return confidence intervals for coefficients
from a `spline_mixture()` object
}
