% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_ele.R
\name{coxph_ele}
\alias{coxph_ele}
\title{Fit a CoxPH Model Assuming Exchangeable Linkage Errors}
\usage{
coxph_ele(
  formula,
  data,
  m.rate,
  blocks,
  control = list(init.beta = "default"),
  ...
)
}
\arguments{
\item{formula}{a formula object for the outcome model, the response should be
provided using the `Surv` function and the covariates should be separated by + signs.}

\item{data}{a data.frame with linked data used in "formula" and "formula.m" (optional)}

\item{m.rate}{block-wise mismatch rates (should be a vector with length equal
to the number of blocks) - by default assume a single block.}

\item{blocks}{block indicators.}

\item{control}{an optional list variable to of control arguments including
"init.beta" for the initial outcome model coefficient estimates) - by
default is the naive estimator.}

\item{...}{the option to directly pass "control" arguments}
}
\value{
a list of results from the function called depending on the "family"
specified.
\item{coefficients}{the outcome model coefficient estimates}
\item{standard.errors}{the estimated standard errors}
\item{call}{the matched call}
\item{vcov}{the variance-covariance matrix}
\item{linear.predictors}{the linear predictors}
}
\description{
Fit Cox proportional hazards regression adjusted for mismatched data based
on the approach developed in Vo et al., 2024 assuming exchangeable linkage
errors. Block-wise mismatch rates are assumed to be known.
}
\note{
The references below discuss the implemented framework in more detail.
}
\references{
Vo, T. H., Garès, V., Zhang, L. C., Happe, A., Oger, E.,
Paquelet, S., & Chauvet, G. (2024). Cox regression with linked data.
Statistics in Medicine, 43(2), 296-314.\cr
}
