% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_mixture.R
\name{coxph_mixture}
\alias{coxph_mixture}
\title{Fit a CoxPH Model Using a Mixture-Modeling Approach}
\usage{
coxph_mixture(
  formula,
  data,
  m.formula,
  safe.matches,
  m.rate,
  control = list(init.beta = "default", init.gamma = "default", fy = "default", max.iter
    = 1000, tol = 1e-04, cmax.iter = 1000),
  ...
)
}
\arguments{
\item{formula}{a formula object for the outcome model, the response should be
provided using the `Surv` function and the covariates should be separated by + signs.}

\item{data}{a data.frame with linked data used in "formula" and "formula.m" (optional)}

\item{m.formula}{a one-sided formula object for the mismatch indicator model, with the
covariates on the right of "~". The default is an intercept-only model corresponding
to a constant mismatch rate)}

\item{safe.matches}{an indicator variable for safe matches (TRUE : record can be treated as a
correct match and FALSE : record may be mismatched). The default is FALSE for all matches.}

\item{m.rate}{the assumed overall mismatch rate (a proportion between 0 and 1). If
not provided, no overall mismatch rate is assumed.}

\item{control}{an optional list variable to customize the initial parameter estimates
("init.beta" for the outcome model and "init.gamma" for the mismatch indicator model),
estimated marginal density of the response ("fy"), maximum iterations for the
EM algorithm ("max.iter"), maximum iterations for the subroutine in the constrained
logistic regression function ("cmax.iter"), and convergence tolerance for
the termination of the EM algorithm ("tol").}

\item{...}{the option to directly pass "control" arguments}
}
\value{
a list of results from the function called depending on the "family"
specified.
\item{coefficients}{the outcome model coefficient estimates}
\item{match.prob}{the posterior correct match probabilities for observations
given parameter estimates}
\item{objective}{a variable that tracks the negative log pseudo-likelihood
for all iterations of the EM algorithm.}
\item{standard.errors}{the estimated standard errors}
\item{m.coefficients}{the correct match model coefficient estimates}
\item{call}{the matched call}
\item{wfit}{an internal-use object for the predict function}
\item{vcov}{the variance-covariance matrix}
\item{Lambdahat_0}{the baseline cumulative hazard (using weighted Breslow estimator) when the family is "cox"}
\item{g_Lambdahat_0}{the baseline cumulative hazard for the marginal density
of the response variable (using Nelson-Aalen estimator)}
}
\description{
Fit Cox proportional hazards regression adjusted for mismatched data.
Information about the underlying record linkage process can be
incorporated into the method if available (e.g., assumed overall mismatch rate,
safe matches, predictors of match status, or predicted probabilities of correct
matches).
}
\note{
The reference below discuss the implemented framework in more detail. The standard
errors are estimated Louis' method for the "cox" family (Bukke et al., 2023).\cr
}
\references{
Bukke, P., Ben-David, E., Diao, G., Slawski, M., & West, B. T. (2025).
Cox Proportional Hazards Regression Using Linked Data: An Approach Based on Mixture Modelling.\cr
}
