% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable_mixture.R
\name{ctable_mixture}
\alias{ctable_mixture}
\title{Contingency Table Analysis Based on Mixture-Modeling}
\usage{
ctable_mixture(
  formula,
  data = NULL,
  m.rate,
  control = list(max.iter = 100, tol = 1e-04),
  ...
)
}
\arguments{
\item{formula}{a formula object with the left and right hand sides specifying the column
and row variable of the flat table, respectively.}

\item{data}{a data frame, list, environment, or contingency table (see `ftable.formula`).}

\item{m.rate}{the assumed mismatch rate. It should be greater than 0 and less than 1.}

\item{control}{an optional list variable for the EM algorithm to customize the maximum
iterations ("max.iter") and convergence tolerance for termination ("tol").}

\item{...}{further arguments to the default ftable method may also be passed as arguments,
see `ftable.formula` and `ftable.default`. Also, control settings may be passed directly
instead of through the control argument.}
}
\value{
a list of results.
\item{phat}{a matrix of the estimated cell probabilities.}
\item{phat0}{a matrix of the estimated cell probabilities for mismatched pairs under
an independence model. The corresponding parameter is a nuisance parameter.}
\item{vcov_phat}{the variance-covariance matrix of phat.}
\item{ftable}{a flat contingency table in matrix format containing the estimated counts
of each combination of the variables’ levels.}
\item{objs}{the sequence of objective values for the non-negative likelihood, which is
expected to be decreasing.}
\item{m.rate}{the mismatch rate assumed.}
\item{call}{the matched call.}
}
\description{
Estimate the cell probabilities p_ij underlying a two-way contingency table under a multinomial sampling scheme
(i.e., the total number of counts is considered fixed). Note that in the absence of mismatches, this corresponds
to fitting a saturated model (the estimated cell probabilities will be exactly equal to the corresponding relative
frequencies). This estimator will be consistent under regularity conditions for correctly matched data. In the
presence of mismatches, however, the relative frequencies in the table will be biased. An adjustment is made by
fitting a mixture between a saturated model and an independence model. The mixture proportions for these models are
"1 – m.rate" and "m.rate", respectively. The mismatch rate cannot be estimated but needs to be pre-specified.
}
\note{
The reference below discusses the underlying methodology in more detail.
}
\references{
Slawski, M., West, B. T., Bukke, P., Diao, G., Wang, Z., & Ben-David, E. (2023).
A General Framework for Regression with Mismatched Data Based on Mixture Modeling.
Under Review. < \doi{10.48550/arXiv.2306.00909} >\cr
}
