% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ele.R
\name{glm_ele}
\alias{glm_ele}
\title{Fit a GLM Assuming Exchangeable Linkage Errors}
\usage{
glm_ele(
  formula,
  data = NULL,
  family = "gaussian",
  m.rate,
  blocks,
  weight.matrix = "all",
  control = list(init.beta = "default"),
  ...
)
}
\arguments{
\item{formula}{a formula object for the outcome model, with the covariate(s) on
the right of "~" and the response on the left.}

\item{data}{a data.frame with linked data used in "formula" and "formula.m" (optional)}

\item{family}{the type of regression model ("gaussian" - default, "poisson",
"binomial", "gamma"). Standard link functions are used ("identity" for Gaussian,
"log" for Poisson and Gamma, and "logit" for binomial).}

\item{m.rate}{block-wise mismatch rates (should be a vector with length equal
to the number of blocks) - by default assume a single block.}

\item{blocks}{block indicators.}

\item{weight.matrix}{the type of weight matrix ("ratio-type",
"Lahiri-Larsen", "BLUE", or "all" (default))}

\item{control}{an optional list variable to of control arguments including
"init.beta" for the initial outcome model coefficient estimates) - by
default is the naive estimator when the weight matrix is ratio-type or
Lahiri-Larsen and is the Lahiri-Larsen estimator for the BLUE weight matrix.}

\item{...}{the option to directly pass "control" arguments}
}
\value{
a list of results from the function called depending on the "family" specified.
\item{coefficients}{the outcome model coefficient estimates}
\item{family}{the type of (outcome) regression model}
\item{standard.errors}{the estimated standard errors}
\item{call}{the matched call}
\item{dispersion}{the dispersion parameter estimate}
\item{vcov}{the variance-covariance matrix}
\item{linear.predictors}{the linear predictors}
\item{df.residual}{the residual degrees of freedom}
}
\description{
Fit generalized linear regression models adjusted for mismatched data according
to the ELE approach (Chambers, 2009).
The function currently supports Gaussian, gamma, Poisson, and logistic (binary)
models. Block-wise mismatch rates are assumed to be known.
}
\note{
The reference below discusses the implemented framework in more detail.
}
\references{
Chambers, R. (2009). Regression analysis of probability-linked data.\cr
}
