% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_mixture.R
\name{glm_mixture}
\alias{glm_mixture}
\title{Fit a GLM Using a Mixture-Modeling Approach}
\usage{
glm_mixture(
  formula,
  data,
  family = c("gaussian", "poisson", "binomial", "gamma"),
  m.formula,
  safe.matches,
  m.rate,
  control = list(init.beta = "default", init.gamma = "default", fy = "default", max.iter
    = 1000, tol = 1e-04, cmax.iter = 1000),
  ...
)
}
\arguments{
\item{formula}{a formula object for the outcome model, with the covariate(s) on
the right of "~" and the response on the left.}

\item{data}{a data.frame with linked data used in "formula" and "formula.m" (optional)}

\item{family}{the type of regression model ("gaussian" - default, "poisson",
"binomial", "gamma"). Standard link functions are used ("identity" for Gaussian,
"log" for Poisson and Gamma, and "logit" for binomial).}

\item{m.formula}{a one-sided formula object for the mismatch indicator model, with the
covariates on the right of "~". The default is an intercept-only model corresponding
to a constant mismatch rate)}

\item{safe.matches}{an indicator variable for safe matches (TRUE : record can be treated as a
correct match and FALSE : record may be mismatched). The default is FALSE for all matches.}

\item{m.rate}{the assumed overall mismatch rate (a proportion between 0 and 1). If
not provided, no overall mismatch rate is assumed.}

\item{control}{an optional list variable to customize the initial parameter estimates
("init.beta" for the outcome model and "init.gamma" for the mismatch indicator model),
estimated marginal density of the response ("fy"), maximum iterations for the
EM algorithm ("max.iter"), maximum iterations for the subroutine in the constrained
logistic regression function ("cmax.iter"), and convergence tolerance for
the termination of the EM algorithm ("tol").}

\item{...}{the option to directly pass "control" arguments}
}
\value{
a list of results from the function called depending on the "family" specified.
\item{coefficients}{the outcome model coefficient estimates}
\item{match.prob}{the posterior correct match probabilities for observations given parameter estimates}
\item{objective}{a variable that tracks the negative log pseudo-likelihood for all iterations of the EM algorithm.}
\item{family}{the type of (outcome) regression model}
\item{standard.errors}{the estimated standard errors}
\item{m.coefficients}{the correct match model coefficient estimates}
\item{call}{the matched call}
\item{wfit}{an internal-use object for the predict function}
\item{dispersion}{the dispersion parameter estimate}
\item{vcov}{the variance-covariance matrix}
}
\description{
Fit generalized linear regression models adjusted for mismatched data.
The function currently supports Gaussian, gamma, Poisson, and logistic (binary)
models. Information about the underlying record linkage process can be
incorporated into the method if available (e.g., assumed overall mismatch rate,
safe matches, predictors of match status, or predicted probabilities of correct
matches).
}
\note{
The references below discuss the implemented framework in more detail. The standard
errors are estimated using the sandwich formula (Slawski et al., 2023).\cr
}
\examples{
## commonness score of first and last names used for linkage
mformula <- ~commf + comml
## hand-linked records are considered "safe" matches
safematches <- ifelse(lifem$hndlnk =="Hand-Linked At Some Level", TRUE, FALSE)
## overall mismatch rate in the data set is assumed to be ~ 0.05
mrate <- 0.05

fit <- fit_mixture(age_at_death ~ poly(unit_yob, 3, raw = TRUE), data = lifem,
                   family = "gaussian", mformula, safematches, mrate)

}
\references{
Slawski, M., West, B. T., Bukke, P., Diao, G., Wang, Z., & Ben-David, E. (2023).
A General Framework for Regression with Mismatched Data Based on Mixture Modeling.
Under Review. < \doi{10.48550/arXiv.2306.00909} >\cr

Slawski, M., Diao, G., Ben-David, E. (2021). A pseudo-likelihood approach to linear
regression with partially shuffled data. Journal of Computational and Graphical
Statistics. 30(4), 991-1003 < \doi{10.1080/10618600.2020.1870482} >
}
