% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers_bayes.R
\name{is_func}
\alias{is_func}
\title{Detect Stan function injections in a priors list}
\usage{
is_func(stan_function)
}
\arguments{
\item{stan_function}{an object to test.}
}
\value{
\code{TRUE} if \code{stan_function} inherits from
\code{"stan_function_string"}, \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if an element was created by \code{stan_func()} and should
be injected into the Stan \code{functions} block.
}
\keyword{internal}
