% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.coxph_ele.R
\name{predict.coxph_ele}
\alias{predict.coxph_ele}
\title{Predictions from a `coxph_ele` Object}
\usage{
\method{predict}{coxph_ele}(object, newdata, type = c("lp", "risk"))
}
\arguments{
\item{object}{the result of a call to `coxph_ele()`}

\item{newdata}{optional new data to obtain predictions for. The original data
is used by default.}

\item{type}{For the "cox" family, the choices are the linear predictor ("lp"),
and the risk score exp(lp) ("risk").}

\item{...}{for future predict arguments}
}
\value{
a vector of predictions based on arguments specified.
}
\description{
Obtain predictions from a `coxph_ele()` object.
}
