% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.coxph_mixture.R
\name{predict.coxph_mixture}
\alias{predict.coxph_mixture}
\title{Predictions from a `coxph_mixture` Object}
\usage{
\method{predict}{coxph_mixture}(
  object,
  newdata,
  type = c("lp", "risk", "expected", "terms", "survival"),
  na.action = na.pass,
  terms = NULL,
  collapse = NULL,
  reference = c("strata", "sample", "zero"),
  ...
)
}
\arguments{
\item{object}{the result of a call to `coxph_mixture()`}

\item{newdata}{optional new data to obtain predictions for. The original data
is used by default.}

\item{type}{For the "cox" family, the choices are the linear predictor ("lp"),
the risk score exp(lp) ("risk"), the expected number of events given the
covariates and follow-up time ("expected"), and the terms of the linear
predictor ("terms"). The survival probability for a subject is equal to
exp(-expected).}

\item{na.action}{a function for what to do with missing values in `newdata`.
The default is to predict "NA".}

\item{terms}{the terms when type = "terms". By default, all terms are included.}

\item{collapse}{optional vector of subject identifiers. If specified, the
output will contain one entry per subject rather than one entry per observation.}

\item{reference}{Reference for centering predictions.
Available options are c("strata" - default, "sample", "zero").}

\item{...}{for future predict arguments}
}
\value{
a vector or matrix of predictions based on arguments specified.
}
\description{
Obtain predictions from a `coxph_mixture()` object using
`predict.coxph()`.
}
