% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.glm_mixture.R
\name{predict.glm_mixture}
\alias{predict.glm_mixture}
\title{Predictions from a `glm_mixture` Object}
\usage{
\method{predict}{glm_mixture}(
  object,
  newdata = NULL,
  type = c("link", "response"),
  se.fit = FALSE,
  dispersion = NULL,
  interval = c("none", "confidence"),
  level = 0.95,
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{the result of a call to `glm_mixture()`}

\item{newdata}{optional new data to obtain predictions for. The original data is used by default.}

\item{type}{the type of prediction. The choices are predictions on the scale
of the linear predictors ("link") or response ("response").}

\item{se.fit}{an indicator of whether standard errors are required (TRUE or FALSE)}

\item{dispersion}{the dispersion to be assumed when computing standard errors.
By default, it is the value returned in the fit object.}

\item{interval}{type of interval calculation.}

\item{level}{required confidence level if interval is set to "confidence".}

\item{na.action}{a function for what to do with missing values in `newdata`.
The default is to predict "NA".}

\item{...}{for future predict arguments}
}
\value{
a vector or matrix of predictions based on arguments specified.
}
\description{
Obtain predictions from a `glm_mixture` object.
}
\examples{
## commonness score of first and last names used for linkage
mformula <- ~commf + comml
## hand-linked records are considered "safe" matches
safematches <- ifelse(lifem$hndlnk =="Hand-Linked At Some Level", TRUE, FALSE)
## overall mismatch rate in the data set is assumed to be ~ 0.05
mrate <- 0.05
fit <- glm_mixture(age_at_death ~ poly(unit_yob, 3, raw = TRUE), data = lifem,
                   family = "gaussian", mformula, safematches, mrate)

predict(fit)

}
