% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.spline_mixture.R
\name{predict.spline_mixture}
\alias{predict.spline_mixture}
\title{Predictions from a `spline_mixture` Object}
\usage{
\method{predict}{spline_mixture}(
  x,
  newdata,
  se.fit = FALSE,
  interval = c("none", "confidence"),
  level = 0.95,
  ...
)
}
\arguments{
\item{x}{the result of a call to `spline_mixture()`}

\item{...}{for additional predict arguments}
}
\value{
matrix of predictions w/ se and point-wise confidence bands
}
\description{
Return predicted values from an `spline_mixture()` object
}
