% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.coxph_ele.R
\name{print.coxph_ele}
\alias{print.coxph_ele}
\title{Print a `coxph_ele` Object}
\usage{
\method{print}{coxph_ele}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{the result of a call to `coxph_ele()`}

\item{digits}{the number of significant digits to print}

\item{...}{for additional print arguments}
}
\value{
invisibly returns the `coxph_ele()` object that is provided as an argument
}
\description{
Print call and outcome model coefficients from a `coxph_ele()` object
}
