% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.coxph_mixture.R
\name{print.coxph_mixture}
\alias{print.coxph_mixture}
\title{Print a `coxph_mixture` Object}
\usage{
\method{print}{coxph_mixture}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{the result of a call to `coxph_mixture()`}

\item{digits}{the number of significant digits to print}

\item{...}{for additional print arguments}
}
\value{
invisibly returns the `coxph_mixture()` object that is provided as an argument
}
\description{
Print call and outcome model coefficients from a `coxph_mixture()` object
}
