% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.glm_ele.R
\name{print.glm_ele}
\alias{print.glm_ele}
\title{Print a `glm_ele` Object}
\usage{
\method{print}{glm_ele}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{the result of a call to `glm_ele()`}

\item{digits}{the number of significant digits to print}

\item{...}{for additional print arguments}
}
\value{
invisibly returns the `glm_ele()` object that is provided as an argument
print(fit)
}
\description{
Print call and outcome model coefficients from a `glm_ele()` object
}
