% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.glm_mixtureBayes.R
\name{print.glm_mixtureBayes}
\alias{print.glm_mixtureBayes}
\title{Print a brief summary of a glm_mixtureBayes model}
\usage{
\method{print}{glm_mixtureBayes}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{glm_mixtureBayes}.}

\item{digits}{Minimum number of significant digits to show.}

\item{...}{Further arguments (unused).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a brief summary of a glm_mixtureBayes model
}
