% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_glm_methods.R
\name{print.glmELE}
\alias{print.glmELE}
\title{Print a `glmELE` Object}
\usage{
\method{print}{glmELE}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"glmELE"}.}

\item{digits}{The number of significant digits to print. Defaults to
\code{max(3L, getOption("digits") - 3L)}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Prints the function call and the estimated coefficient matrices from a fitted
`glmELE` object.
}
