% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_glm_methods.R
\name{print.glmMixture}
\alias{print.glmMixture}
\title{Print a glmMixture Object}
\usage{
\method{print}{glmMixture}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{glmMixture}.}

\item{digits}{The number of significant digits to use.}

\item{...}{Additional arguments.}
}
\description{
Print a glmMixture Object
}
