% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spline_mixture.R
\name{print.spline_mixture}
\alias{print.spline_mixture}
\title{Print an `spline_mixture` Object}
\usage{
\method{print}{spline_mixture}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{the result of a call to `spline_mixture()`}

\item{digits}{the number of significant digits to print}

\item{...}{for additional print arguments}
}
\value{
invisibly returns the `spling_mixture()` object
that is provided as an argument
}
\description{
Print call and input for a `spline_mixture()` object
}
