% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_coxph_methods.R
\name{print.summary.coxphELE}
\alias{print.summary.coxphELE}
\title{Print Summary of a coxphELE Object}
\usage{
\method{print}{summary.coxphELE}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{summary.coxphELE}.}

\item{digits}{The number of significant digits to use.}

\item{signif.stars}{Logical; if \code{TRUE}, 'significance stars' are printed.}

\item{...}{Additional arguments passed to \code{printCoefmat}.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Prints the summary object generated by \code{\link{summary.coxphELE}}.
}
