% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_ctable_methods.R
\name{print.summary.ctableMixture}
\alias{print.summary.ctableMixture}
\title{Print Summary of Adjusted Contingency Table}
\usage{
\method{print}{summary.ctableMixture}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{summary.ctableMixture}.}

\item{digits}{Integer; number of significant digits to use. Defaults to 3.}

\item{...}{Additional arguments passed to \code{\link{print.default}}.}
}
\value{
The argument \code{x}, invisibly.
}
\description{
Prints the summary generated by \code{\link{summary.ctableMixture}}.
}
