% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_glm_methods.R
\name{print.summary.glmELE}
\alias{print.summary.glmELE}
\title{Print Summary of a \code{glmELE} Object}
\usage{
\method{print}{summary.glmELE}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"summary.glmELE"}.}

\item{digits}{The number of significant digits to print.}

\item{signif.stars}{Logical; if \code{TRUE}, significance stars are printed.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Prints the summary of a \code{glmELE} fit.
}
