% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{print.summary.glmMixBayes}
\alias{print.summary.glmMixBayes}
\title{Print method for summary.glmMixBayes}
\usage{
\method{print}{summary.glmMixBayes}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"summary.glmMixBayes"}.}

\item{digits}{Significant digits to use in printing.}

\item{signif.stars}{Logical; if \code{TRUE}, print significance stars for coefficients.
Defaults to \code{getOption("show.signif.stars")}.}

\item{...}{Additional arguments (unused).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print method for summary.glmMixBayes
}
