% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.surv_mixtureBayes.R
\name{print.summary.surv_mixtureBayes}
\alias{print.summary.surv_mixtureBayes}
\title{Print a `summary.surv_mixtureBayes` Object}
\usage{
\method{print}{summary.surv_mixtureBayes}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a \code{summary.surv_mixtureBayes} object.}

\item{digits}{the number of significant digits to print.}

\item{...}{for additional print arguments (unused).}
}
\value{
invisibly returns the \code{summary.surv_mixtureBayes} object that is provided.
}
\description{
Print a formatted summary for a \code{surv_mixtureBayes} fit, including
posterior mean and uncertainty summaries for both mixture components.
}
