% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_survreg_methods.R
\name{print.summary.survMixBayes}
\alias{print.summary.survMixBayes}
\title{Print a summary.survMixBayes object}
\usage{
\method{print}{summary.survMixBayes}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.survMixBayes}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Additional arguments (unused).}
}
\value{
The object \code{x} (invisibly).
}
\description{
Print a summary.survMixBayes object
}
