% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.surv_mixtureBayes.R
\name{print.surv_mixtureBayes}
\alias{print.surv_mixtureBayes}
\title{Print a `surv_mixtureBayes` Object}
\usage{
\method{print}{surv_mixtureBayes}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{surv_mixtureBayes}.}

\item{digits}{the number of significant digits to print.}

\item{...}{for additional print arguments (unused).}
}
\value{
invisibly returns the \code{surv_mixtureBayes} object that is provided.
}
\description{
Print the model call and posterior mean coefficients for component 1.
}
