% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_mixture.R
\name{spline_mixture}
\alias{spline_mixture}
\title{Spline least squares regression Based on Mixture Modeling}
\usage{
spline_mixture(
  formula,
  data,
  control = list(max.iter = 100, fy = "default"),
  ...
)
}
\arguments{
\item{formula}{model formula of the form y ~ s(x), where x and y are
continuous variables contained in "data". This formula will be passed on
to calls of functions in the mgcv package, and can be used to specify
the number of knots.}

\item{data}{underlying data set}

\item{control}{an optional list variable with additional control arguments.}

\item{...}{other arguments to be passed to the gam() call in mgcv.}
}
\value{
a list of results having the following components.
\item{B}{matrix of spline basis function evaluations from mgcv. Necessary for computing fitted values.}
\item{par_coef}{the parameters of the Normal distribution for the spline coefficients, i.e., its mean and covariance matrix.}
\item{par_sigmasq}{the parameters of the Inverse-Gamma distribution. (shape and rate)of the error variance sigmasq}
\item{par_tausq}{the parameters of the Inverse-Gamma distribution (shape and rate) of the variance of the (conditional) prior for the spline coefficients.}
\item{par_alpha}{the parameters of the Beta distribution for the mismatch rate alpha.}
\item{par_m}{the parameters of the Bernoulli distributions of the mismatch indicator (posterior probability of mismatch)}
}
\description{
This function performs least squares spline fitting accounting
for mismatched (x,y) pairs. Spline functionality is imported
from the {mgcv} package. Fitting is based on the Bayesian
formulation of the mixture model as described in Slawski et al. (2024).
Parameters are estimated using a vanilla mean-field variational approximation
of the posterior. The updates can be performed in closed form.
Currently only a constant mismatch rate model is supported.
}
