% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.coxph_ele.R
\name{summary.coxph_ele}
\alias{summary.coxph_ele}
\title{Summarize a `coxph_ele` Object}
\usage{
\method{summary}{coxph_ele}(object, ...)
}
\arguments{
\item{object}{the result of a call to `coxph_ele()`}

\item{...}{for additional summary arguments}
}
\value{
a list of results from the function called depending on the "family" specified.
\item{call}{the matched call}
\item{coefficients}{a matrix with the outcome model's coefficient estimates, standard errors, t or z values, and p-values}#'
}
\description{
Summarize results from a `coxph_ele()` object
}
