% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.coxph_mixture.R
\name{summary.coxph_mixture}
\alias{summary.coxph_mixture}
\title{Summarize a `coxph_mixture` Object}
\usage{
\method{summary}{coxph_mixture}(object, ...)
}
\arguments{
\item{object}{the result of a call to `coxph_mixture()`}

\item{...}{for additional summary arguments}
}
\value{
a list of results from the function called depending on the "family" specified.
\item{call}{the matched call}
\item{family}{the assumed type of (outcome) regression model}
\item{coefficients}{a matrix with the outcome model's coefficient estimates, standard errors, t or z values, and p-values}
\item{m.coefficients}{a matrix with the correct match model's coefficient estimates and standard errors}
\item{avgcmr}{the average correct match rate among all records}
\item{match.prob}{the posterior correct match probabilities for observations given parameter estimates}
}
\description{
Summarize results from a `coxph_mixture()` object
}
