% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glm_ele.R
\name{summary.glm_ele}
\alias{summary.glm_ele}
\title{Summarize a "glm_ele" Object}
\usage{
\method{summary}{glm_ele}(object, ...)
}
\arguments{
\item{object}{the result of a call to `glm_ele()`}

\item{...}{for additional summary arguments}
}
\value{
a list of results from the function called depending on the "family" specified.
\item{call}{the matched call}
\item{family}{the assumed type of (outcome) regression model}
\item{coefficients}{a matrix with the outcome model's coefficient estimates, standard errors, t or z values, and p-values}
\item{dispersion}{the dispersion parameter estimate when the family is a Generalized Linear Model}
}
\description{
Summarize results from a `glm_ele()` object
}
