% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glm_mixtureBayes.R
\name{summary.glm_mixtureBayes}
\alias{summary.glm_mixtureBayes}
\title{Summarize a glm_mixtureBayes model fit}
\usage{
\method{summary}{glm_mixtureBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{glm_mixtureBayes}.}

\item{...}{Not used.}
}
\value{
An object of class \code{"summary.glm_mixtureBayes"}, which is printed with a custom method.
}
\description{
Summarize a glm_mixtureBayes model fit
}
