% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{summary.glmMixBayes}
\alias{summary.glmMixBayes}
\title{Summary Method for glmMixBayes Models}
\usage{
\method{summary}{glmMixBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{glmMixBayes}.}

\item{...}{Not used.}
}
\value{
An object of class \code{"summary.glmMixBayes"}, which is printed with a custom method.
}
\description{
Summary Method for glmMixBayes Models
}
\examples{
\donttest{
data(lifem)

# lifem data preprocessing
# For computational efficiency in the example, we work with a subset of the lifem data.
lifem <- lifem[order(-(lifem$commf + lifem$comml)), ]
lifem_small <- rbind(
  head(subset(lifem, hndlnk == 1), 100),
  head(subset(lifem, hndlnk == 0), 20)
)

x <- cbind(1, poly(lifem_small$unit_yob, 3, raw = TRUE))
y <- lifem_small$age_at_death

adj <- adjMixBayes(
  linked.data = lifem_small,
  priors = list(theta = "beta(2, 2)")
)

fit <- plglm(
  age_at_death ~ poly(unit_yob, 3, raw = TRUE),
  family = "gaussian",
  adjustment = adj,
  control = list(
    iterations = 200,
    burnin.iterations = 100,
    seed = 123
  )
)

summary(fit)
}

}
