% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spline_mixture.R
\name{summary.spline_mixture}
\alias{summary.spline_mixture}
\title{Summarize an `spline_mixture` Object}
\usage{
\method{summary}{spline_mixture}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{the result of a call to `spline_mixture()`}

\item{digits}{the number of significant digits to print}

\item{...}{for additional summary arguments}
}
\value{
invisibly returns the `spline_mixture()` object
that is provided as an argument
}
\description{
Summarize results from a `spline_mixture()` object
}
