% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.surv_mixtureBayes.R
\name{summary.surv_mixtureBayes}
\alias{summary.surv_mixtureBayes}
\title{Summarize a surv_mixtureBayes model fit}
\usage{
\method{summary}{surv_mixtureBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{surv_mixtureBayes}.}

\item{...}{Not used.}
}
\value{
An object of class \code{"summary.surv_mixtureBayes"}, which is printed with a custom method.
}
\description{
Summarize a surv_mixtureBayes model fit
}
