% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.coxph_ele.R
\name{vcov.coxph_ele}
\alias{vcov.coxph_ele}
\title{Variance-Covariance Matrix from a `coxph_ele` Object}
\usage{
\method{vcov}{coxph_ele}(object, ...)
}
\arguments{
\item{object}{an object of class "coxph_ele".}

\item{...}{for additional `vcov` arguments.}
}
\value{
Return the variance-covariance matrix for the model parameter estimates.
}
\description{
Extract the variance-covariance matrix in the `coxph_ele` object for
the model parameter estimates.
}
