% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.coxph_mixture.R
\name{vcov.coxph_mixture}
\alias{vcov.coxph_mixture}
\title{Variance-Covariance Matrix from a `coxph_mixture` Object}
\usage{
\method{vcov}{coxph_mixture}(object, ...)
}
\arguments{
\item{object}{an object of class "coxph_mixture".}

\item{...}{for additional `vcov` arguments.}
}
\value{
Return the variance-covariance matrix for the model parameter estimates.
}
\description{
Extract the variance-covariance matrix in the `coxph_mixture` object for
the model parameter estimates.
}
