% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.glm_ele.R
\name{vcov.glm_ele}
\alias{vcov.glm_ele}
\title{Variance-Covariance Matrix from a `glm_ele` Object}
\usage{
\method{vcov}{glm_ele}(object, ...)
}
\arguments{
\item{object}{an object of class "glm_ele".}

\item{...}{for additional `vcov` arguments.}
}
\value{
Return the variance-covariance matrix for the model parameter estimates.
}
\description{
Extract the variance-covariance matrix in the `glm_ele` object for
the model parameter estimates.
}
