% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.glm_mixture.R
\name{vcov.glm_mixture}
\alias{vcov.glm_mixture}
\title{Variance-Covariance Matrix from a `glm_mixture` Object}
\usage{
\method{vcov}{glm_mixture}(object, ...)
}
\arguments{
\item{object}{an object of class `glm_mixture`.}

\item{...}{for additional `vcov` arguments.}
}
\value{
Return the variance-covariance matrix for the model parameter estimates.
}
\description{
Extract the variance-covariance matrix in the `glm_mixture` object for
the model parameter estimates.
}
