% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.glm_mixtureBayes.R
\name{vcov.glm_mixtureBayes}
\alias{vcov.glm_mixtureBayes}
\title{Covariance matrix of coefficient estimates for glm_mixtureBayes}
\usage{
\method{vcov}{glm_mixtureBayes}(object, ...)
}
\arguments{
\item{object}{A \code{glm_mixtureBayes} model object.}

\item{...}{Not used.}
}
\value{
Posterior covariance matrix of component 1's coefficient vector.
}
\description{
Covariance matrix of coefficient estimates for glm_mixtureBayes
}
