% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_glm_methods.R
\name{vcov.glmELE}
\alias{vcov.glmELE}
\title{Extract Variance-Covariance Matrix from a `glmELE` Object}
\usage{
\method{vcov}{glmELE}(object, weight.matrix = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"glmELE"}.}

\item{weight.matrix}{Character string specifying which weighting method to return.
Defaults to the first method found in the object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A matrix of the estimated covariances between the parameter estimates.
}
\description{
Extracts the variance-covariance matrix of the main parameters for a specific 
weighting method.
}
