% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.spline_mixture.R
\name{vcov.spline_mixture}
\alias{vcov.spline_mixture}
\title{Variance-Covariance Matrix from a `spline_mixture` Object}
\usage{
\method{vcov}{spline_mixture}(x, ...)
}
\arguments{
\item{x}{the result of a call to `spline_mixture()`}

\item{...}{for additional vcov arguments}
}
\value{
variance-covariance matrix
}
\description{
Extract the variance-covariance matrix
}
