% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.surv_mixtureBayes.R
\name{vcov.surv_mixtureBayes}
\alias{vcov.surv_mixtureBayes}
\title{Covariance matrix of coefficient estimates for surv_mixtureBayes}
\usage{
\method{vcov}{surv_mixtureBayes}(object, ...)
}
\arguments{
\item{object}{A \code{surv_mixtureBayes} model object.}

\item{...}{Not used.}
}
\value{
Posterior covariance matrix of component 1's coefficient vector.
}
\description{
Covariance matrix of coefficient estimates for surv_mixtureBayes
}
