% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_adjMixBayes.R
\name{adjMixBayes}
\alias{adjMixBayes}
\title{Constructor for Secondary Analysis of Linked Data Based on Bayesian Mixture Modeling}
\usage{
adjMixBayes(linked.data = NULL)
}
\arguments{
\item{linked.data}{A data.frame containing the linked dataset.}
}
\value{
An object of class \code{c("adjMixBayes", "adjustment")}. To minimize
memory overhead, the underlying \code{linked.data} is stored by reference
within an environment inside this object.
}
\description{
Specifies the linked data and information on the underlying record linkage
process for a mismatch error adjustment using a Bayesian framework based on
mixture modeling as developed by Gutman et al. (2016). This framework uses
a mixture model for pairs of linked records whose two components reflect distributions
conditional on match status, i.e., correct match or mismatch.
Posterior inference is carried out via data augmentation or multiple
imputation.
}
\details{
Explicit provision of \code{linked.data} is strongly recommended for
reproducibility and to ensure the adjustment object fully encapsulates
the necessary data for downstream model fitting.

#' @examples
}
\note{
The reference below discusses the implemented framework in more detail.
}
\section{Example: Using the included brfss demonstration dataset}{
data(brfss, package = "postlink")

adj_object <- adjMixBayes(linked.data = brfss)
}

\references{
Gutman, R., Sammartino, C., Green, T., & Montague, B. (2016). Error adjustments for file
linking methods using encrypted unique client identifier (eUCI) with application to recently
released prisoners who are HIV+. \emph{Statistics in Medicine}, 35(1), 115–129. \doi{10.1002/sim.6586}
}
