% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{coef.glmELE}
\alias{coef.glmELE}
\title{Extract Coefficients for glmELE}
\usage{
\method{coef}{glmELE}(object, weight.matrix = NULL, ...)
}
\arguments{
\item{object}{A fitted glmELE object.}

\item{weight.matrix}{Character specifying which weighting method to return.}

\item{...}{Additional arguments.}
}
\description{
Extract Coefficients for glmELE
}
