% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{coef.survMixBayes}
\alias{coef.survMixBayes}
\title{Extract Coefficients for survMixBayes}
\usage{
\method{coef}{survMixBayes}(object, ...)
}
\arguments{
\item{object}{A fitted survMixBayes object.}

\item{...}{Additional arguments.}
}
\description{
Extract Coefficients for survMixBayes
}
