% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.ctablemixture.R
\name{confint.ctablemixture}
\alias{confint.ctablemixture}
\title{Confidence intervals from a `ctable_mixture` Object.}
\usage{
\method{confint}{ctablemixture}(object, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class "ctablemixture".}

\item{level}{the required confidence level.}

\item{...}{for additional `confint` arguments.}
}
\value{
a data frame with the upper and lower confidence limits for each
cell probability. These limits are labelled as (1-level)/2 and
1 - (1-level)/2 (e.g., 2.5% and 97.5% for level 0.95).
}
\description{
Compute confidence intervals for the cell probabilities from the fitted model.
}
