% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.spline_mixture.R
\name{confint.spline_mixture}
\alias{confint.spline_mixture}
\title{Confidence Intervals from a `spline_mixture` Object}
\usage{
\method{confint}{spline_mixture}(x, level = 0.95)
}
\arguments{
\item{x}{the result of a call to `spline_mixture()`}

\item{level}{the desired confidence level (default is 0.95)}

\item{...}{for additional confint arguments}
}
\value{
a matrix with the upper and lower confidence limits for each
outcome model coefficient. These limits are labelled as (1-level)/2 and
1 - (1-level)/2 (e.g., 2.5\% and 97.5\% for level 0.95).
}
\description{
Return confidence intervals for coefficients
from a `spline_mixture()` object
}
