% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_coxph.R
\name{coxphELE}
\alias{coxphELE}
\title{Fit a CoxPH Model Assuming Exchangeable Linkage Errors}
\usage{
coxphELE(
  x,
  y,
  cens,
  m.rate,
  blocks,
  audit.size = NULL,
  control = list(init.beta = NULL),
  ...
)
}
\arguments{
\item{x}{A matrix or data.frame of covariates (design matrix).}

\item{y}{A numeric vector of observed time-to-event outcomes.}

\item{cens}{A numeric vector indicating censoring status (1 = censored, 0 = event).
Note: This is the reverse of the standard `Surv` object convention where 1 usually
indicates an event.}

\item{m.rate}{block-wise mismatch rates (should be a vector with length equal
to the number of blocks) - by default assume a single block.}

\item{blocks}{block indicators.}

\item{audit.size}{a vector of block sizes in the audit sample (selected by
simple random sampling) if used to estimate the m.rate (optional). If a
single value is provided, assume the same value for all blocks and put out a warning.}

\item{control}{an optional list variable to of control arguments including
"init.beta" for the initial outcome model coefficient estimates) - by
default is the naive estimator.}

\item{...}{the option to directly pass "control" arguments}
}
\value{
a list of results from the function called depending on the "family"
specified.
\item{coefficients}{the outcome model coefficient estimates}
\item{var}{the variance-covariance matrix}
\item{linear.predictors}{the linear predictors}
\item{means}{Column means of the covariate matrix `x`.}
\item{n}{Number of observations.}
\item{nevent}{Number of events.}
}
\description{
Fit Cox proportional hazards regression adjusted for mismatched data based
on the approach developed in Vo et al., 2024 assuming exchangeable linkage
errors. Block-wise mismatch rates are assumed to be known.
}
\references{
Vo, T. H., Garès, V., Zhang, L. C., Happe, A., Oger, E.,
Paquelet, S., & Chauvet, G. (2024). Cox regression with linked data.
Statistics in Medicine, 43(2), 296-314.\cr
}
