% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{df.residual.plglm}
\alias{df.residual.plglm}
\title{Extract Residual Degrees of Freedom for plglm models}
\usage{
\method{df.residual}{plglm}(object, ...)
}
\arguments{
\item{object}{A fitted plglm object.}

\item{...}{Additional arguments.}
}
\description{
Extract Residual Degrees of Freedom for plglm models
}
