% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{mi_with}
\alias{mi_with}
\title{Posterior allocation based pooling generic}
\usage{
mi_with(object, ...)
}
\arguments{
\item{object}{A fitted Bayesian mixture model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A pooled model object.
}
\description{
Generic function for posterior allocation based pooling of Bayesian
mixture model fits.
}
