% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_survreg_methods.R
\name{mi_with.survMixBayes}
\alias{mi_with.survMixBayes}
\title{Pooling method for survMixBayes objects}
\usage{
\method{mi_with}{survMixBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{survMixBayes}.}

\item{...}{Additional arguments (unused).}
}
\value{
A pooled object of class \code{mi_link_pool_survreg}.
}
\description{
Performs posterior allocation based pooling for a \code{survMixBayes} fit.
}
