% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plctable.R
\name{plctable}
\alias{plctable}
\title{Analysis of Contingency Tables with Linkage Error Adjustment}
\usage{
plctable(
  formula,
  adjustment,
  subset,
  na.action,
  exclude = c(NA, NaN),
  control = list(),
  ...
)
}
\arguments{
\item{formula}{a formula object with the left and right hand sides specifying the column
and row variable of the flat table, respectively.}

\item{adjustment}{An object inheriting from class \code{"adjustment"}, or a
\code{list} containing the necessary parameter specifications.}

\item{subset}{An optional vector specifying a subset of observations to be used.}

\item{na.action}{A function which indicates what should happen when the data contain NAs.}

\item{exclude}{Vector of values to be excluded when forming the table (passed to \code{xtabs}).}

\item{control}{A list of parameters for controlling the linkage error
adjustment process.}

\item{...}{Additional arguments passed to the internal engine.}
}
\value{
A fitted model object tailored for contingency table analysis.
}
\description{
\code{plctable} constructs a contingency table from linked data and fits a model
adjusting for linkage errors.
}
