% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.glm_ele.R
\name{predict.glm_ele}
\alias{predict.glm_ele}
\title{Predictions From a `glm_ele` Object}
\usage{
\method{predict}{glm_ele}(object, newdata, type = c("response", "link"))
}
\arguments{
\item{object}{the result of a call to `glm_ele()`}

\item{newdata}{optional new data to obtain predictions for. The original data is used by default.}

\item{type}{the type of prediction. The choices are predictions on the scale
of the linear predictors ("link") or response ("response").}

\item{...}{for future predict arguments}
}
\value{
a vector of predictions based on arguments specified.
}
\description{
Obtain predictions from a `glm_ele()` object.
}
