% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{print.adjustment}
\alias{print.adjustment}
\title{Print Base Method for Adjustment Objects}
\usage{
\method{print}{adjustment}(x, ...)
}
\arguments{
\item{x}{An object of class "adjustment".}

\item{...}{Additional arguments.}
}
\description{
Print Base Method for Adjustment Objects
}
