% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_coxph_methods.R
\name{print.coxphELE}
\alias{print.coxphELE}
\title{Print a coxphELE Object}
\usage{
\method{print}{coxphELE}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{coxphELE}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Additional arguments passed to \code{\link{print.default}}.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Prints a short summary of the fitted CoxPH model with linkage error adjustment,
displaying the call (if available) and the estimated coefficients.
}
