% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_coxph_methods.R
\name{print.coxphMixture}
\alias{print.coxphMixture}
\title{Print a coxphMixture Object}
\usage{
\method{print}{coxphMixture}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{coxphMixture}.}

\item{digits}{The number of significant digits to use.}

\item{...}{Additional arguments.}
}
\description{
Print a coxphMixture Object
}
