% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ctablemixture.R
\name{print.ctablemixture}
\alias{print.ctablemixture}
\title{Print a `ctable_mixture` Object}
\usage{
\method{print}{ctablemixture}(object, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{an object of class "ctablemixture".}

\item{digits}{the number of significant digits to use for printing.}

\item{...}{for additional `print` arguments.}
}
\value{
Invisibly returns the object from `ctable_mixture()`
that is provided as an argument.
}
\description{
Print the call and input data for the `ctable_mixture` object.
}
