% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{print.glmMixBayes}
\alias{print.glmMixBayes}
\title{Print a brief summary of a glmMixBayes model}
\usage{
\method{print}{glmMixBayes}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{glmMixBayes}.}

\item{digits}{Minimum number of significant digits to show.}

\item{...}{Further arguments (unused).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a brief summary of a glmMixBayes model
}
