% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_glm_methods.R
\name{print.glmMixture}
\alias{print.glmMixture}
\title{Print a glmMixture Object}
\usage{
\method{print}{glmMixture}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{glmMixture}.}

\item{digits}{The number of significant digits to use.}

\item{...}{Additional arguments.}
}
\description{
Print a glmMixture Object
}
\examples{
# Load the LIFE-M demo dataset
data(lifem)

# Phase 1: Adjustment Specification
# We model the correct match indicator via logistic regression using
# name commonness scores (commf, comml) and a 5\% expected mismatch rate.
adj_object <- adjMixture(
 linked.data = lifem,
 m.formula = ~ commf + comml,
 m.rate = 0.05,
 safe.matches = hndlnk
)

# Phase 2: Estimation & Inference
# Fit a Gaussian regression model utilizing a cubic polynomial for year of birth.
fit <- plglm(
 age_at_death ~ poly(unit_yob, 3, raw = TRUE),
 family = "gaussian",
 adjustment = adj_object
)

print(fit)

}
