% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.glm_mixtureBayes.R
\name{print.mi_link_pool_glm}
\alias{print.mi_link_pool_glm}
\title{Print a pooled MI object from `glm_mixtureBayes`}
\usage{
\method{print}{mi_link_pool_glm}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{an object of class \code{mi_link_pool_glm}.}

\item{digits}{the number of significant digits to print.}

\item{...}{for additional print arguments.}
}
\value{
invisibly returns the \code{mi_link_pool_glm} object that is provided.
}
\description{
Print pooled estimates (Rubin's rules) computed from multiple imputations
induced by posterior linkage draws.
}
