% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{print.mi_link_pool_glm}
\alias{print.mi_link_pool_glm}
\title{Print a pooled MI object from glmMixBayes}
\usage{
\method{print}{mi_link_pool_glm}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{An object of class \code{mi_link_pool_glm}.}

\item{digits}{the number of significant digits to print.}

\item{...}{further arguments (unused).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a pooled MI object from glmMixBayes
}
