% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{print.mi_link_pool_glm}
\alias{print.mi_link_pool_glm}
\title{Print pooled regression results}
\usage{
\method{print}{mi_link_pool_glm}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{An object of class \code{mi_link_pool_glm}, typically returned by
\code{mi_with()} for a \code{glmMixBayes} fit.}

\item{digits}{the number of significant digits to print.}

\item{...}{further arguments (unused).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print pooled regression results
}
\examples{
\dontrun{
data(lifem)

# lifem data preprocessing
# For computational efficiency in the example, we work with a subset of the lifem data.
lifem <- lifem[order(-(lifem$commf + lifem$comml)), ]
lifem_small <- rbind(
  head(subset(lifem, hndlnk == 1), 100),
  head(subset(lifem, hndlnk == 0), 20)
)

x <- cbind(1, poly(lifem_small$unit_yob, 3, raw = TRUE))
y <- lifem_small$age_at_death

fit <- glmMixBayes(
  X = x,
  y = y,
  family = "gaussian",
  control = list(
    iterations = 200,
    burnin.iterations = 100,
    seed = 123
  )
)

pooled_fit <- mi_with(
  object = fit,
  data = lifem_small,
  formula = age_at_death ~ poly(unit_yob, 3, raw = TRUE),
  family = gaussian()
)

print(pooled_fit, digits = 4)
}

}
