% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_survreg_methods.R
\name{print.mi_link_pool_survreg}
\alias{print.mi_link_pool_survreg}
\title{Print method for pooled survreg MI object}
\usage{
\method{print}{mi_link_pool_survreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mi_link_pool_survreg}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for pooled survreg MI object
}
