% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_survreg_methods.R
\name{print.mi_link_pool_survreg}
\alias{print.mi_link_pool_survreg}
\title{Print pooled Cox regression results}
\usage{
\method{print}{mi_link_pool_survreg}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{An object of class \code{mi_link_pool_survreg}, typically returned by
\code{mi_with()} for a \code{survMixBayes} fit.}

\item{digits}{the number of significant digits to print.}

\item{...}{further arguments (unused).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print pooled Cox regression results
}
\examples{
\dontrun{
set.seed(301)
n <- 150
trt <- rbinom(n, 1, 0.5)

# Component 1 represents correct links (signal),
# and component 2 represents incorrect links (noise).
Z_true <- 2 - rbinom(n, 1, 0.8)

time1 <- rweibull(n, shape = 1.5, scale = exp(1 + 0.8 * trt))  # Correct links
time2 <- rweibull(n, shape = 1.2, scale = exp(1 + 0.2 * trt))  # Incorrect links
obs_time <- ifelse(Z_true == 1, time1, time2)

cens_time <- rexp(n, rate = 0.1)
status <- as.integer(obs_time <= cens_time)
obs_time <- pmin(obs_time, cens_time)

linked_df <- data.frame(time = obs_time, status = status, trt = trt)

adj <- adjMixBayes(linked.data = linked_df)

fit <- plsurvreg(
  survival::Surv(time, status) ~ trt,
  dist = "weibull",
  adjustment = adj,
  control = list(iterations = 200, burnin.iterations = 100, seed = 123)
)

dat <- data.frame(
 time = y[, "time"],
 event = y[, "event"],
 X
 )

pooled_obj <- mi_with(
  object = fit,
  data = linked_df,
  formula = survival::Surv(time, status) ~ trt
)

print(pooled_obj, digits = 4)
}

}
