% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_coxph_methods.R
\name{summary.coxphMixture}
\alias{summary.coxphMixture}
\title{Summarizing Cox PH Mixture Fits}
\usage{
\method{summary}{coxphMixture}(object, conf.int = 0.95, scale = 1, ...)
}
\arguments{
\item{object}{An object of class \code{coxphMixture}.}

\item{conf.int}{The confidence level for the confidence intervals of the hazard ratios.}

\item{scale}{Scale factor for the standard errors (default is 1).}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{summary.coxphMixture} containing:
\item{call}{The function call.}
\item{n}{Total number of observations.}
\item{nevent}{Number of events.}
\item{coefficients}{Matrix of coefficients for the outcome model.}
\item{m.coefficients}{Matrix of coefficients for the mismatch model.}
\item{conf.int}{Matrix of confidence intervals for the hazard ratios.}
\item{logtest}{Log-likelihood information (Outcome Model).}
\item{avgcmr}{The average posterior probability of a correct match.}
}
\description{
\code{summary} method for class \code{coxphMixture}. Provides a detailed summary
of the fitted model, including coefficients, hazard ratios, standard errors,
z-statistics, and p-values for both the outcome model and the mismatch model.
}
