% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ctablemixture.R
\name{summary.ctablemixture}
\alias{summary.ctablemixture}
\title{Summarize a `ctable_mixture` Object}
\usage{
\method{summary}{ctablemixture}(object, ...)
}
\arguments{
\item{object}{an object of class "ctablemixture".}

\item{...}{for additional `summary` arguments.}
}
\value{
a list of results.
\item{call}{the matched call.}
\item{phat}{a matrix of the estimated cell probabilities.}
\item{vcov_phat}{the variance-covariance matrix of phat.}
\item{ftable}{a flat contingency table in matrix format containing the estimated counts
of each combination of the variables’ levels.}
\item{chi2_statistic}{the Pearson’s chi-squared test statistic based on
the `ftable` value}
}
\description{
Summarize the results in the `ctable_mixture` object.
}
