% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{summary.glmMixBayes}
\alias{summary.glmMixBayes}
\title{Summarize a glmMixBayes model fit}
\usage{
\method{summary}{glmMixBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{glmMixBayes}.}

\item{...}{Not used.}
}
\value{
An object of class \code{"summary.glmMixBayes"}, which is printed with a custom method.
}
\description{
Summarize a glmMixBayes model fit
}
