% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_glm_methods.R
\name{summary.glmMixture}
\alias{summary.glmMixture}
\title{Summarizing GLM Mixture Fits}
\usage{
\method{summary}{glmMixture}(object, dispersion = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{glmMixture}.}

\item{dispersion}{The dispersion parameter for the family used. If NULL, it is inferred from object.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{summary.glmMixture} containing:
\item{call}{The component from object.}
\item{family}{The component from object.}
\item{df.residual}{The residual degrees of freedom.}
\item{coefficients}{Matrix of coefficients for the outcome model.}
\item{m.coefficients}{Matrix of coefficients for the mismatch model.}
\item{dispersion}{Estimated dispersion parameter.}
\item{cov.unscaled}{The estimated covariance matrix.}
\item{match.prob}{The posterior match probabilities.}
}
\description{
\code{summary} method for class \code{glmMixture}.
}
\examples{
# Load the LIFE-M demo dataset
data(lifem)

# Phase 1: Adjustment Specification
# We model the correct match indicator via logistic regression using
# name commonness scores (commf, comml) and a 5\% expected mismatch rate.
adj_object <- adjMixture(
 linked.data = lifem,
 m.formula = ~ commf + comml,
 m.rate = 0.05,
 safe.matches = hndlnk
)

# Phase 2: Estimation & Inference
# Fit a Gaussian regression model utilizing a cubic polynomial for year of birth.
fit <- plglm(
 age_at_death ~ poly(unit_yob, 3, raw = TRUE),
 family = "gaussian",
 adjustment = adj_object
)

summary(fit)

}
