% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_glm_methods.R
\name{summary.glmMixture}
\alias{summary.glmMixture}
\title{Summarizing GLM Mixture Fits}
\usage{
\method{summary}{glmMixture}(object, dispersion = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{glmMixture}.}

\item{dispersion}{The dispersion parameter for the family used. If NULL, it is inferred from object.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{summary.glmMixture} containing:
\item{call}{The component from object.}
\item{family}{The component from object.}
\item{deviance}{The component from object.}
\item{coefficients}{Matrix of coefficients for the outcome model.}
\item{m.coefficients}{Matrix of coefficients for the mismatch model.}
\item{dispersion}{Estimated dispersion parameter.}
\item{df.residual}{The residual degrees of freedom.}
\item{cov.scaled}{The estimated covariance matrix.}
\item{resid.summary}{The residual summary statistics.}
}
\description{
\code{summary} method for class \code{glmMixture}.
}
