% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ele_coxph_methods.R
\name{vcov.coxphELE}
\alias{vcov.coxphELE}
\title{Variance-Covariance Matrix for coxphELE Objects}
\usage{
\method{vcov}{coxphELE}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxphELE}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A matrix of the estimated covariances between the parameter estimates.
}
\description{
Extracts the variance-covariance matrix of the main parameters (outcome model coefficients)
from a fitted \code{coxphELE} model.
}
