% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.ctablemixture.R
\name{vcov.ctablemixture}
\alias{vcov.ctablemixture}
\title{Variance-Covariance Matrix from a `ctable_mixture` Object}
\usage{
\method{vcov}{ctablemixture}(object, ...)
}
\arguments{
\item{object}{an object of class "ctablemixture".}

\item{...}{for additional `vcov` arguments.}
}
\value{
A matrix of the estimated covariances between the cell probabilities.
This should have row and column names corresponding to the row and column
variables’ levels of the cells in the underlying two-way contingency table.
}
\description{
Extract the variance-covariance matrix for the estimated cell probabilities
in the `ctable_mixture` object.
}
