% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_glm_methods.R
\name{vcov.glmMixBayes}
\alias{vcov.glmMixBayes}
\title{Covariance matrix of coefficient estimates for glmMixBayes}
\usage{
\method{vcov}{glmMixBayes}(object, ...)
}
\arguments{
\item{object}{A \code{glmMixBayes} model object.}

\item{...}{Not used.}
}
\value{
Posterior covariance matrix of component 1's coefficient vector.
}
\description{
Covariance matrix of coefficient estimates for glmMixBayes
}
