% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixbayes_survreg_methods.R
\name{vcov.survMixBayes}
\alias{vcov.survMixBayes}
\title{Variance-covariance for survMixBayes coefficients}
\usage{
\method{vcov}{survMixBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{survMixBayes}.}

\item{...}{Additional arguments (unused).}
}
\value{
A covariance matrix.
}
\description{
Returns an empirical posterior covariance matrix of component 1 regression coefficients.
}
